+++ CHANGES.rst
@@ -3,6 +3,7 @@ Changes
 
 0.14.15 (unreleased)
 
+- Do not crash on empty XML file (@ekandler)
 
 
 0.14.14 (2020-04-25)