+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changes
 0.14.15 (unreleased)
 
 - Do not crash on empty XML file (@ekandler)
+- Option to print timestamp in ISO8601 (@kitlaan)
 - Option to not print LastChange subscription variable (@kitlaan)