+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes
 - Do not crash on empty XML file (@ekandler)
 - Option to print timestamp in ISO8601 (@kitlaan)
 - Option to not print LastChange subscription variable (@kitlaan)
+- Test with Python 3.8 (@scop)
 
 
 0.14.14 (2020-04-25)